app/pyproject.toml
[project]
name = "roc-portal"
version = "0.1.0"
description = "ROC Portal Power BI - Secure embed portal with RBAC and multi-tenancy"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "psycopg[binary]>=3.2.0",
    "redis>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.18",
    "jinja2>=3.1.4",
    "itsdangerous>=2.2.0",
    "httpx>=0.28.0",
    "msal>=1.31.0",
    "structlog>=24.4.0",
    "babel>=2.16.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.0",
    "email-validator>=2.2.0",
    "aiosmtplib>=3.0.2",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
    "types-passlib",
    "types-python-jose",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "S", "RUF"]
ignore = ["S101"]  # allow assert in tests

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
